+++ llm_assistant/tests/__init__.py
@@ -4,3 +4,4 @@ from . import test_pause
 from . import test_llm_retry
 from . import test_stream_trace_sink
 from . import test_stream_duration_cap
+from . import test_effective_tools